Cleaning up the structure of the project by moving code to components.
