+++ components/Shop/shop.css
@@ -1,5 +1,3 @@
-@import "./product/product.css";
-
 .shop {
   background-color: rgb(241, 237, 237);
   display: flex;
