+++ components/styles.css
@@ -1,1 +1,4 @@
 @import "./Add-form/add-form.css";
+@import "./Product/product.css";
+@import "./Shop/shop.css";
+@import "./Sidebar/sidebar.css";
