+++ services/index.js
@@ -1,7 +1,1 @@
-export const productsDataPath = "../data/products.json";
-export const categoriesDataPath = "../data/categories.json";
-export const archivesDataPath = "../data/archives.json";
-export const commentsDataPath = "../data/comments.json";
-export const postsDataPath = "../data/posts.json";
-
 export const baseUrl = "http://localhost:3000/";
